+++ .github/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## v0.2.1
+
+1. Added documentation
+1. Bug fixes and stability improvements
+
+## v0.2.0
+
+1. Bugfixes and code improvements
+1. Added search functionality
+1. Added interviews to Dashboard calendar
+1. Added toasts when adding, editing or deleting data
+
 ## v0.1.0
 
 1. First release! :tada: